+++ packages/eslint-config-react/.eslintrc.js
@@ -12,5 +12,8 @@ module.exports = {
     'no-magic-numbers': ['off'], // since defining a bunch of consts for lint rules is cumbersome
     'quote-props': ['error', 'consistent-as-needed'], // since the majority of rules are in quotes with little outlyers;
     'max-lines': ['off'], // We're not interested in breaking up rule files
-  }
+  },
+  parserOptions: {
+    requireConfigFile: false,
+  },
 }