+++ packages/eslint-config/.eslintrc.js
@@ -10,7 +10,10 @@ module.exports = {
   extends: './index.js',
   rules: {
     'no-magic-numbers': ['off'], // since defining a bunch of consts for lint rules is cumbersome
-    'quote-props': ['error', 'consistent-as-needed'], // since the majority of rules are in quotes with little outlyers;
+    'quote-props': ['error', 'consistent-as-needed'], // since the majority of rules are in quotes with little outliers;
     'max-lines': ['off'], // We're not interested in breaking up rule files
-  }
+  },
+  parserOptions: {
+    requireConfigFile: false,
+  },
 }