+++ packages/eslint-config/core/bestpractices.js
@@ -61,6 +61,12 @@ module.exports = {
 
 
     /**
+     * enforce default clauses in switch statements to be last
+     */
+    'default-case-last': ['error'],
+
+
+    /**
      * enforce default parameters to be last
      */
     'default-param-last': ['off'],
@@ -82,7 +88,7 @@ module.exports = {
      * require the use of === and !==
      */
     'eqeqeq': ['error', 'always', {
-      null: 'never',
+      null: 'always',
     }],
 
 
@@ -273,6 +279,7 @@ module.exports = {
         1000, // for time math
       ],
       enforceConst: true,
+      ignoreDefaultValues: true,
     }],
 
 
@@ -330,14 +337,15 @@ module.exports = {
         '^desc(riptor)?', // for ESNext decorators
         '^req(uest)?', // for Express requests
         '^res(ponse)?', // for Express responses
+        '^draft(State)?', // for immer producers
       ],
       ignorePropertyModificationsFor: [
         'Component', // for React HOCs
         'connection', // alternative name for Koa ctx
         'ctx', // for Koa routing
-        'draftState', // for immer producers
         'target', // for ESNext decorators
         'client', // for Websocket servers
+        'event', // for general events objects.
       ],
     }],