+++ packages/eslint-config/util/importExtensions.js
@@ -1,0 +1,7 @@
+module.exports = [
+  '.js',
+  '.jsx',
+  '.mjs',
+  '.cjs',
+  '.json',
+]